Got it working with the as-lib
